backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pdfplumber==0.10.4
python-docx==0.8.11
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
aiofiles==23.2.1
cors==1.0.1
